chore: updated `FirebaseInitScript`
