+++ .changeset/angry-lamps-help.md
@@ -1,0 +1,5 @@
+---
+"twitch-clip": patch
+---
+
+Fixed performance metrics initialization.
