--- src/firebase/client/client.tsx
@@ -1,22 +1,0 @@
-// Import the functions you need from the SDKs you need
-import { initializeApp } from "firebase/app"
-
-// TODO: Add SDKs for Firebase products that you want to use
-// https://firebase.google.com/docs/web/setup#available-libraries
-
-// Your web app's Firebase configuration
-// For Firebase JS SDK v7.20.0 and later, measurementId is optional
-const firebaseConfig = {
-  apiKey: process.env.NEXT_PUBLIC_FIREBASE_API_KEY,
-  appId: process.env.NEXT_PUBLIC_APPID,
-  authDomain: process.env.NEXT_PUBLIC_AUTHDOMAIN,
-  measurementId: process.env.NEXT_PUBLIC_MEASUREMENTID,
-  messagingSenderId: process.env.NEXT_PUBLIC_MESSAGINGSENDERID,
-  projectId: process.env.NEXT_PUBLIC_PROJECTID,
-  storageBucket: process.env.NEXT_PUBLIC_STORAGEBUCKET,
-}
-
-// Initialize Firebase
-const clientApp = initializeApp(firebaseConfig)
-
-export { clientApp }
